% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_tests.R
\name{diagnostic_tests}
\alias{diagnostic_tests}
\title{Diagnostic tests of a dynamic linear model (independence, homoskedasticity and normality)}
\usage{
diagnostic_tests(dlm_filtered, n_lags)
}
\arguments{
\item{dlm_filtered}{The output from \code{dlm::dlmFilter()}.}

\item{n_lags}{}
}
\value{
A data.frame/tibble of the model results including a column specifying whether the assumptions of independence, homoskedasticity and normality of the residuals is satisfied.
}
\description{
This function implements the Ljung-Box test of independence, the H-test of homoskedasticity and the Jarque-Bera test of normality on the model residuals.
}
\examples{
Soon!
}
