% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_dlm_viz.R
\name{dlm_diag_viz}
\alias{dlm_diag_viz}
\title{Visualizations for model diagnostics}
\usage{
dlm_diag_viz(dlm_mod)
}
\arguments{
\item{dlm_mod}{The output of the \code{dlm::full_dlm_modeling} function.}
}
\value{
A \code{list} of \code{ggplot} objects.
}
\description{
The function makes some visualizations for determining the model's goodness of it: residuals (raw), residuals (standardized), autocorrelation function (ACF), QQ-plot and the p-values of the Ljung-Box statistic.
}
\examples{
print("Soon!")
}
