% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_dlm_viz.R
\name{dlm_state_viz}
\alias{dlm_state_viz}
\title{Visualizations of DLM model estimates}
\usage{
dlm_state_viz(dlm_mod)
}
\arguments{
\item{dlm_mod}{The output of the \code{dlm::full_dlm_modeling} function.}
}
\value{
A \code{list} of \code{ggplot} objects.
}
\description{
The function visualizes the model's state components.
}
\examples{
print("Soon!")
}
