% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_tests.R
\name{dt_homoskedasticity}
\alias{dt_homoskedasticity}
\title{H-test of homoskedasticity of the model residuals}
\usage{
dt_homoskedasticity(dlm_filtered)
}
\arguments{
\item{dlm_filtered}{The output from \code{dlm::dlmFilter()}.}
}
\value{
A data.frame/tibble of the model results including a column specifying whether the assumption of homoskedasticity of the residuals is satisfied.
}
\description{
The H-test procedure implements a test against the null hypothesis of homoskedastic residuals.
}
\examples{
print("Soon!")
}
