% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_tests.R
\name{dt_independence}
\alias{dt_independence}
\title{Ljung-Box test of independence of the model residuals}
\usage{
dt_independence(dlm_mod, n_lags)
}
\arguments{
\item{dlm_mod}{The output of the \code{dlm::full_dlm_modeling} function.}

\item{n_lags}{The number of lags to consider in the test.}
}
\value{
A data.frame/tibble of the model results including a column specifying whether the assumption of independence of the residuals is satisfied.
}
\description{
The Ljung-Box procedure implements a test against the null hypothesis of independence.
}
\examples{
print("Soon!")
}
