% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_tests.R
\name{dt_normality}
\alias{dt_normality}
\title{Jarque-Bera test of normality of the model residuals}
\usage{
dt_normality(dlm_filtered)
}
\arguments{
\item{dlm_filtered}{The output from \code{dlm::dlmFilter()}.}
}
\value{
A data.frame/tibble of the model results including a column specifying whether the assumption of normality of the residuals is satisfied.
}
\description{
Jarque-Bera test of normality of the model residuals
}
\examples{
print("Soon!")
}
