% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_viz.R
\name{full_dlm_viz}
\alias{full_dlm_viz}
\title{Visualizations for model diagnostics}
\usage{
full_dlm_viz(dlm_mod)
}
\arguments{
\item{dlm_mod}{The output of the \code{dlm::full_dlm_modeling} function.}
}
\value{
A \code{list} of \code{ggplot} objects.
}
\description{
The function makes some visualizations for determining the model's goodness of it: residuals, autocorrelation function (ACF), QQ-plot and the p-values of the Ljung-Box statistic.
}
\examples{
Soon!
}
