% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_likelihood-and-aic.R
\name{get_AIC}
\alias{get_AIC}
\title{Calculate the value of the AIC}
\usage{
get_AIC(log_likelihood, n_state_vars, n_hyper_params)
}
\arguments{
\item{log_likelihood}{Numeric vector of length 1. Value of the log-likelihood function as provided by the \code{get_log_likelihood()} function.}

\item{n_state_vars}{Numeric value of length 1. Number of state variables in the dynamic linear model.}

\item{n_hyper_params}{Numeric value of length 1. Number of stochastic hyperparameters in the dynamic linear model.}
}
\value{
The Akaike Information Criterion (AIC) of the model.
}
\description{
Calculate the value of the AIC
}
\examples{
Soon!
}
