% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_likelihood-and-aic.R
\name{get_log_likelihood}
\alias{get_log_likelihood}
\title{Calculate the value of the log-likelihood function}
\usage{
get_log_likelihood(dlm_filtered)
}
\arguments{
\item{dlm_filtered}{The output from \code{dlm::dlmFilter()}.}
}
\value{
The value of the log-likelihood function.
}
\description{
Calculate the value of the log-likelihood function
}
\examples{
Soon!
}
