% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ukksi}
\alias{ukksi}
\title{UK Killed or Seriously Injured (KSI)}
\format{
A data frame/tibble with 192 rows and 2 columns:
\describe{
\item{date}{Date class}
\item{value}{Numeric class}
}
}
\source{
\url{https://github.com/sinhrks/stan-statespace}
}
\usage{
ukksi
}
\description{
A dataset containing the number of killed or seriously injured (KSI) in the UK from January 1969 to December 1984.
}
\keyword{datasets}
